Avances en los modelos

+++ Models/PersonaModel.cs
@@ -5,9 +5,9 @@
         public int id { get; }
         public string nombre { get; }
         public string direccion { get; }
-        public int telefono { get; }
+        public long telefono { get; }
 
-        public PersonaModel(int id, string nombre, string direccion, int telefono)
+        public PersonaModel(int id, string nombre, string direccion, long telefono)
         {
             this.id = id;
             this.nombre = nombre;
